WIP: code cleanup

+++ build.sh
@@ -32,25 +32,27 @@ do
   echo "====== (esm)COMPILING: \$(npm bin)/tsc -p ${SRCDIR}/tsconfig-es2015.json ====="
   $(npm bin)/tsc -p ${SRCDIR}/tsconfig-es2015.json
 
-  echo "======      BUNDLING: ${SRCDIR} ====="
-  (
-    cd  ${SRCDIR}
-    echo "..."  # here just to have grep match something and not exit with 1
-    ../../rollup-test/node_modules/.bin/rollup -c rollup.config.js
-  ) 2>&1 | grep -v "as external dependency"
+  if (true); then
+    echo "======      BUNDLING: ${SRCDIR} ====="
+    (
+      cd  ${SRCDIR}
+      echo "..."  # here just to have grep match something and not exit with 1
+      ../../../node_modules/.bin/rollup -c rollup.config.js
+    ) 2>&1 | grep -v "as external dependency"
 
-  # workaround for https://github.com/rollup/rollup/issues/626
-  sed -i '' "s/ class exports\./ class /g" ${DESTDIR}/esm/${PACKAGE}.umd.js
+    # workaround for https://github.com/rollup/rollup/issues/626
+    sed -i '' "s/ class exports\./ class /g" ${DESTDIR}/esm/${PACKAGE}.umd.js
 
-  $(npm bin)/tsc  \
-      --out ${UMDES5PATH} \
-      --target es5 \
-      --allowJs \
-      ${UMDES6PATH} \
-      modules/\@angular/manual_typings/globals.d.ts \
-      modules/\@angular/typings/es6-collections/es6-collections.d.ts \
-      modules/\@angular/typings/es6-promise/es6-promise.d.ts
-  rm ${UMDES6PATH}
+    $(npm bin)/tsc  \
+        --out ${UMDES5PATH} \
+        --target es5 \
+        --allowJs \
+        ${UMDES6PATH} \
+        modules/\@angular/manual_typings/globals.d.ts \
+        modules/\@angular/typings/es6-collections/es6-collections.d.ts \
+        modules/\@angular/typings/es6-promise/es6-promise.d.ts
+    rm ${UMDES6PATH}
+  fi
 done
 
 TSCONFIG=./modules/@angular/tsconfig.json
